Keep fixing errors
Similar to what happened to the store of Blocks, rounds and events also
need to be handled as `int64` everywhere.

This is WIP, build is still broken. Going step by step.

+++ src/node/graph.go
@@ -83,7 +83,7 @@ func (g *Graph) GetParticipantEvents() map[string]map[string]poset.Event {
 func (g *Graph) GetRounds() []poset.RoundInfo {
 	res := []poset.RoundInfo{}
 
-	round := 0
+	round := int64(0)
 
 	store := g.Node.core.poset.Store
